fix: use message box instead of dialog

+++ embarker/embarker/preferencesmanager.py
@@ -4,7 +4,6 @@ from embarker import preferences
 from embarker.autosave import get_documents_folder
 import embarker.commands as ebc
 
-
 class PreferencesWindow(QtWidgets.QWidget):
 
     def __init__(self, parent=None):
@@ -186,30 +185,10 @@ class UserColorWidget(QtWidgets.QWidget):
         self.refresh_color()
 
 
-class ResetPreferences(QtWidgets.QDialog):
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Reset Preferences")
-        label = QtWidgets.QLabel("Do you want to reset all your preferences ?")
-
-        self.reset_button = QtWidgets.QPushButton("Reset")
-        self.cancel_button = QtWidgets.QPushButton("Cancel")
-
-        self.reset_button.clicked.connect(lambda _: self.reset_preferences())
-        self.cancel_button.clicked.connect(lambda _: self.reject())
-
-        button_layout = QtWidgets.QHBoxLayout()
-        button_layout.addWidget(self.reset_button)
-        button_layout.addWidget(self.cancel_button)
-        layout = QtWidgets.QVBoxLayout()
-        layout.addWidget(label)
-        layout.addLayout(button_layout)
-        self.setLayout(layout)
-
-    def reset_preferences(self):
+def reset_preferences(_):
+    r =  QtWidgets.QMessageBox.question(
+        None,
+        'Reset preferences ?',
+        'This action is not undoable.\nWould you like to continue?')
+    if r == QtWidgets.QMessageBox.Yes:
         preferences.delete_all()
-        self.accept()
-
-    def sizeHint(self):
-        return QtCore.QSize(150, 250)
